Put the annotation code where it belongs
Four places, from review.

The two color swatches are ordinary widgets with nothing annotation-specific in
them, so they are ImGui::ColorSwatch() and ImGui::StrokeFillSwatches() in
imgui_ext, beside FlatButton.

app-zoom.cpp is pan, zoom and the coordinate transforms; the annotate tool's
interaction had no business there, and neither did the panel's row helpers in
app-windows.cpp among the statistics and file windows. Both, along with the
marks drawn over a text annotation being typed, are src/app-annotations.cpp --
the feature in one file, as the other app-*.cpp are one concern each.

Code motion and one rename: the union of tokens across the files involved is
unchanged but for the new file's includes and StrokeFillSwatches' name.

The README gains the display server and its vector graphics, which went
unmentioned, and annotations.
