+++ src/app-windows.cpp
@@ -714,566 +714,6 @@ void HDRViewApp::draw_statistics_window()
     }
 }
 
-/// A color swatch of a stated size, opening a picker when clicked.
-/**
-    ColorEdit4's own swatch is always one frame tall, which is too big to pair two of them into a control
-    that still lines up with the widgets beside it.
-*/
-static bool color_swatch(const char *id, float4 &color, float size)
-{
-    bool changed = false;
-    ImGui::PushID(id);
-    if (ImGui::ColorButton("##swatch", ImVec4(color.x, color.y, color.z, color.w), ImGuiColorEditFlags_AlphaPreviewHalf,
-                           ImVec2(size, size)))
-        ImGui::OpenPopup("##picker");
-    if (ImGui::BeginPopup("##picker"))
-    {
-        changed = ImGui::ColorPicker4("##picker4", &color.x, ImGuiColorEditFlags_AlphaBar);
-        ImGui::EndPopup();
-    }
-    ImGui::PopID();
-    return changed;
-}
-
-/// Stroke and fill as one control, the way an illustration tool shows them: the fill in front, the stroke
-/// behind it and drawn as a ring. Which is which is what the overlap says, so neither needs a label.
-static bool stroke_fill_swatches(const char *id, float4 &stroke, float4 &fill)
-{
-    // Two thirds of a frame each, overlapping by half of that, so the pair is exactly one frame tall and
-    // sits on the same line as the widgets beside it.
-    const float  h    = ImGui::GetFrameHeight();
-    const float  sw   = h * 2.f / 3.f;
-    const float  off  = sw * 0.5f;
-    const ImVec2 base = ImGui::GetCursorScreenPos();
-
-    ImGui::PushID(id);
-
-    // Behind, so it is submitted first and allows the fill to take the hover where the two meet.
-    ImGui::SetNextItemAllowOverlap();
-    ImGui::SetCursorScreenPos(ImVec2(base.x + off, base.y + off));
-    bool changed = color_swatch("stroke", stroke, sw);
-    ImGui::SetItemTooltip("Stroke: the outline of a shape, and the color of a text annotation.");
-
-    // Its middle punched out before the fill goes over it, so it reads as an outline, not a second fill.
-    const ImVec2 lo = ImGui::GetItemRectMin(), hi = ImGui::GetItemRectMax();
-    const float  t = (hi.x - lo.x) * 0.3f;
-    ImGui::GetWindowDrawList()->AddRectFilled(ImVec2(lo.x + t, lo.y + t), ImVec2(hi.x - t, hi.y - t),
-                                              ImGui::GetColorU32(ImGuiCol_WindowBg));
-
-    ImGui::SetCursorScreenPos(base);
-    changed |= color_swatch("fill", fill, sw);
-    ImGui::SetItemTooltip("Fill: the inside of a closed shape. Fully transparent leaves it unfilled.");
-
-    ImGui::PopID();
-
-    // Both were placed by hand, so the layout has to be told how much room they took between them.
-    ImGui::SetCursorScreenPos(base);
-    ImGui::Dummy(ImVec2(sw + off, sw + off));
-    return changed;
-}
-
-/// What a row calls an annotation, cut with an ellipsis at \p width.
-/**
-    A text annotation is named by what it says, which can be a paragraph, so the row shows as much of it as
-    it has room for.
-*/
-static std::string row_name(const Annotation &a, float width)
-{
-    std::string name =
-        a.shape == Annotation::Shape::Text && a.label.empty() && !a.text.empty() ? a.text : a.display_label();
-
-    if (ImGui::CalcTextSize(name.c_str()).x <= width)
-        return name;
-
-    const float ellipsis = ImGui::CalcTextSize("...").x;
-    while (!name.empty() && ImGui::CalcTextSize(name.c_str()).x + ellipsis > width) name.pop_back();
-    return name + "...";
-}
-
-/// The icon the panel and the shape picker show for \p shape.
-static const char *annotation_shape_icon(Annotation::Shape shape)
-{
-    switch (shape)
-    {
-    case Annotation::Shape::Rect: return ICON_MY_SHAPE_RECT;
-    case Annotation::Shape::Ellipse: return ICON_MY_SHAPE_ELLIPSE;
-    case Annotation::Shape::Line: return ICON_MY_SHAPE_LINE;
-    case Annotation::Shape::Arrow: return ICON_MY_SHAPE_ARROW;
-    case Annotation::Shape::Text: return ICON_MY_SHAPE_TEXT;
-    case Annotation::Shape::Freehand: return ICON_MY_SHAPE_FREEHAND;
-    default: return ICON_MY_ANNOTATE;
-    }
-}
-
-void HDRViewApp::draw_annotations_window()
-{
-    auto img = current_image();
-    if (!img)
-    {
-        ImGui::TextDisabled("No image loaded.");
-        return;
-    }
-
-    auto     &list   = img->annotations;
-    const int active = active_annotation();
-
-    // Taken before anything below flattens them, for the popup that wants ordinary widgets: the rows put
-    // both to nothing so their buttons sit against each other and stay one line tall.
-    const ImVec2 frame_padding  = ImGui::GetStyle().FramePadding;
-    const float  item_spacing_x = ImGui::GetStyle().ItemSpacing.x;
-
-    ImGui::PushStyleVarY(ImGuiStyleVar_FramePadding, 0.f);
-    ImGui::Checkbox("Show annotations in viewport", &m_draw_annotations);
-    ImGui::PopStyleVar();
-
-    // One row of controls, editing whichever annotation is in hand, or the look the next one will be drawn
-    // with when none is. The same widgets either way, so there is nothing to learn twice.
-    Annotation &edited = active >= 0 ? list[size_t(active)] : m_annotation_style;
-    draw_annotation_controls(edited);
-
-    // The viewport asks for this when a text annotation is placed with a click, which leaves it saying
-    // nothing, and when one is double-clicked to be edited. Both open its row for typing.
-    if (m_annotation_edit_text && active >= 0)
-    {
-        m_annotation_renaming   = active;
-        m_annotation_rename_was = list[size_t(active)].text;
-        snprintf(m_annotation_rename, sizeof(m_annotation_rename), "%s", m_annotation_rename_was.c_str());
-    }
-
-    // Applied after the list is drawn: removing a row mid-list renumbers the ones still to come.
-    int erase = -1;
-
-    // Where the rows are, so a drag can say which one the cursor is over.
-    float first_row_top = 0.f, row_height = 0.f;
-
-    // The same table the Images panel's list is: an outer border, striped rows, and rows reaching the
-    // full width. A child window would inset them by its padding.
-    static constexpr ImGuiTableFlags table_flags = ImGuiTableFlags_NoSavedSettings | ImGuiTableFlags_SizingFixedFit |
-                                                   ImGuiTableFlags_BordersOuter | ImGuiTableFlags_RowBg |
-                                                   ImGuiTableFlags_ScrollY;
-
-    if (ImGui::BeginTable("AnnotationList", 1, table_flags))
-    {
-        ImGui::TableSetupColumn("", ImGuiTableColumnFlags_WidthStretch);
-
-        // A row is as tall as its text, like the Images panel's, and its buttons sit against each other:
-        // each already carries its own highlight, which is all the separation they need. No frame padding
-        // either, which would otherwise be subtracted from a square button and push its glyph off center.
-        ImGui::PushStyleVar(ImGuiStyleVar_FramePadding, ImVec2(0.f, 0.f));
-        ImGui::PushStyleVarX(ImGuiStyleVar_ItemSpacing, 0.f);
-
-        // Square, and big enough for the widest glyph any of these buttons draws: an icon is wider than the
-        // font size it is drawn at, so a square of that would clip it.
-        float side = ImGui::GetTextLineHeight();
-        for (const char *icon : {ICON_MY_VISIBILITY, ICON_MY_VISIBILITY_OFF, ICON_MY_LOCK, ICON_MY_LOCK_OPEN,
-                                 ICON_MY_SMOOTH, ICON_MY_POLYLINE, ICON_MY_FONT, ICON_MY_TRASH_CAN})
-            side = std::max(side, ImGui::CalcTextSize(icon).x);
-        const ImVec2 icon_sz{side, side};
-
-        auto flat_toggle = [icon_sz](const char *on, const char *off, bool &value, const char *tooltip)
-        {
-            const bool clicked = ImGui::FlatButton(value ? on : off, false, icon_sz);
-            if (clicked)
-                value = !value;
-            ImGui::SetItemTooltip("%s", tooltip);
-            ImGui::SameLine();
-            return clicked;
-        };
-
-        // The renderer's own overlay, listed but never edited: it belongs to the process sending it, and
-        // this is the one place overlays are switched on and off.
-        if (!img->vector_overlay.empty())
-        {
-            ImGui::TableNextRow();
-            ImGui::TableNextColumn();
-            flat_toggle(ICON_MY_VISIBILITY, ICON_MY_VISIBILITY_OFF, img->vector_overlay_visible,
-                        "Draw the renderer's overlay.");
-            ImGui::TextDisabled(ICON_MY_LIST_VIEW " Renderer overlay");
-            ImGui::Tooltip("Drawing commands streamed in by a renderer. They can be hidden here, but only "
-                           "the process sending them can change them.");
-        }
-
-        for (int i = 0; i < int(list.size()); ++i)
-        {
-            auto &a = list[size_t(i)];
-            ImGui::PushID(i);
-
-            ImGui::TableNextRow();
-            ImGui::TableNextColumn();
-
-            const float row_x = ImGui::GetCursorPosX();
-            const float row_w = ImGui::GetContentRegionAvail().x;
-
-            // The highlight goes down first and spans the row, with everything else drawn back over it,
-            // so selecting reads across the whole row.
-            ImGui::SetNextItemAllowOverlap();
-            if (ImGui::Selectable("##row", i == active,
-                                  ImGuiSelectableFlags_AllowOverlap | ImGuiSelectableFlags_SpanAllColumns,
-                                  ImVec2(0.f, icon_sz.y)))
-                set_active_annotation(i);
-
-            // Which row a press took hold of. ImGui's active item cannot say: its id is this row's index,
-            // and reordering moves annotations between indices, so it would stay on the row number and the
-            // list would flicker between two orders.
-            if (ImGui::IsItemActivated())
-                m_annotation_row_drag = i;
-
-            // Taken while the selectable is the item under the cursor, and acted on below, once the name
-            // it applies to has been drawn.
-            const bool renamed_here = ImGui::IsItemHovered() && ImGui::IsMouseDoubleClicked(ImGuiMouseButton_Left);
-
-            // Where the rows sit, for the reorder below. The pitch is measured between the first two rather
-            // than assumed, since a table decides row heights for itself.
-            if (i == 0)
-            {
-                first_row_top = ImGui::GetItemRectMin().y;
-                row_height    = ImGui::GetItemRectSize().y;
-            }
-            else if (i == 1)
-                row_height = ImGui::GetItemRectMin().y - first_row_top;
-
-            ImGui::SameLine(0.f, 0.f);
-            ImGui::SetCursorPosX(row_x);
-
-            flat_toggle(ICON_MY_VISIBILITY, ICON_MY_VISIBILITY_OFF, a.visible, "Draw this annotation.");
-            flat_toggle(ICON_MY_LOCK, ICON_MY_LOCK_OPEN, a.locked,
-                        "A locked annotation cannot be picked up in the viewport.");
-
-            // One column, whatever the shape has to put in it: a scribble runs a curve through its points,
-            // a text annotation picks its face and size, and the rest hold the place so the shape icons
-            // beside them still line up.
-            if (a.shape == Annotation::Shape::Freehand)
-            {
-                if (flat_toggle(ICON_MY_SMOOTH, ICON_MY_POLYLINE, a.smooth,
-                                "Draw this scribble as a curve through its points."))
-                    m_annotation_style.smooth = a.smooth;
-            }
-            else if (a.shape == Annotation::Shape::Text)
-            {
-                if (ImGui::FlatButton(ICON_MY_FONT, false, icon_sz))
-                    ImGui::OpenPopup("##font");
-                ImGui::SetItemTooltip("The face and size this text is drawn in.");
-                ImGui::SameLine();
-                draw_font_popup(a, frame_padding, item_spacing_x);
-            }
-            else
-            {
-                ImGui::Dummy(icon_sz);
-                ImGui::SameLine();
-            }
-
-            if (m_annotation_renaming == i)
-            {
-                // Only the name becomes a field; the icons stay where they are, so a row does not
-                // rearrange itself the moment it is renamed.
-                ImGui::TextUnformatted(annotation_shape_icon(a.shape));
-                ImGui::SameLine();
-
-                // Focused the frame it appears, so the name can be typed straight away; once it holds the
-                // focus it is the active item, so this stops asking for it. Enter or clicking away keeps
-                // the name, and Escape drops it, InputText restoring its own buffer.
-                if (!ImGui::IsAnyItemActive())
-                    ImGui::SetKeyboardFocusHere();
-
-                const float field_w = row_x + row_w - icon_sz.x - ImGui::GetCursorPosX();
-
-                ImGui::SetNextItemWidth(field_w);
-                const bool edited = ImGui::InputTextWithHint("##rename", a.display_label().c_str(), m_annotation_rename,
-                                                             sizeof(m_annotation_rename));
-
-                // So the viewport can find this field's caret and selection and show them on the image.
-                m_annotation_rename_id = ImGui::GetItemID();
-
-                // Focusing a field from code selects all of it; an edit opened from the image continues
-                // what is already there, so the caret goes to the end instead.
-                if (m_annotation_edit_text)
-                    if (auto *state = ImGui::GetInputTextState(m_annotation_rename_id))
-                    {
-                        state->ReloadUserBufAndMoveToEnd();
-                        m_annotation_edit_text = false;
-                    }
-
-                // For a text annotation the row's name is what it says, so this is how the text is typed,
-                // and it is written as it is typed, so the image shows it before the field is left.
-                auto &named = a.shape == Annotation::Shape::Text ? a.text : a.label;
-                if (edited)
-                    named = m_annotation_rename;
-
-                if (ImGui::IsItemDeactivated())
-                {
-                    // Escape asks for what was there before, and every keystroke has already been applied.
-                    if (!ImGui::IsItemDeactivatedAfterEdit())
-                        named = m_annotation_rename_was;
-                    m_annotation_renaming  = -1;
-                    m_annotation_edit_text = false;
-
-                    // A text annotation with nothing to say is not kept: placing one and thinking better
-                    // of it leaves nothing behind, as a click that never became a drag does.
-                    if (a.shape == Annotation::Shape::Text && a.text.empty())
-                        erase = i;
-                }
-            }
-            else
-            {
-                const float name_w = row_x + row_w - icon_sz.x - ImGui::GetCursorPosX() -
-                                     ImGui::CalcTextSize(annotation_shape_icon(a.shape)).x - ImGui::CalcTextSize(" ").x;
-                ImGui::TextUnformatted(
-                    fmt::format("{} {}", annotation_shape_icon(a.shape), row_name(a, name_w)).c_str());
-
-                // Double-clicking the row renames it in place, which is where the name is read.
-                if (renamed_here)
-                {
-                    m_annotation_renaming   = i;
-                    m_annotation_rename_was = a.shape == Annotation::Shape::Text ? a.text : a.label;
-                    snprintf(m_annotation_rename, sizeof(m_annotation_rename), "%s", m_annotation_rename_was.c_str());
-                }
-            }
-
-            ImGui::SameLine(0.f, 0.f);
-            ImGui::SetCursorPosX(row_x + row_w - icon_sz.x);
-            if (ImGui::FlatButton(ICON_MY_TRASH_CAN, false, icon_sz))
-                erase = i;
-            ImGui::SetItemTooltip("Delete this annotation.");
-
-            ImGui::PopID();
-        }
-
-        ImGui::PopStyleVar(2);
-        ImGui::EndTable();
-    }
-
-    if (!ImGui::IsMouseDown(ImGuiMouseButton_Left))
-        m_annotation_row_drag = -1;
-
-    // The row the cursor is over now, which a fast drag can be several away from. The list is drawn in
-    // order, so moving a row is what puts one annotation in front of another.
-    if (erase < 0 && m_annotation_row_drag >= 0 && m_annotation_row_drag < int(list.size()) && row_height > 0.f)
-    {
-        const int to =
-            std::clamp(int((ImGui::GetIO().MousePos.y - first_row_top) / row_height), 0, int(list.size()) - 1);
-        if (to != m_annotation_row_drag)
-        {
-            m_annotation_renaming = -1;
-
-            Annotation moved = list[size_t(m_annotation_row_drag)];
-            list.erase(list.begin() + m_annotation_row_drag);
-            list.insert(list.begin() + to, moved);
-
-            // What is in hand is the annotation, not the row it was in, so the selection follows it.
-            if (active == m_annotation_row_drag)
-                set_active_annotation(to);
-            else if (active > m_annotation_row_drag && active <= to)
-                set_active_annotation(active - 1);
-            else if (active < m_annotation_row_drag && active >= to)
-                set_active_annotation(active + 1);
-
-            m_annotation_row_drag = to;
-        }
-    }
-
-    if (erase >= 0)
-    {
-        // The row being renamed is about to be a different annotation, or none.
-        m_annotation_renaming = -1;
-
-        list.erase(list.begin() + erase);
-        // The rows after it have shifted down, so what was in hand has to shift with them.
-        if (active == erase)
-            set_active_annotation(-1);
-        else if (active > erase)
-            set_active_annotation(active - 1);
-    }
-}
-
-/// The "Add:" label and the picker that says which shape the tool draws next, as one table cell.
-void HDRViewApp::draw_shape_picker(bool named)
-{
-    ImGui::TableNextColumn();
-    const auto &style = ImGui::GetStyle();
-    ImGui::AlignTextToFramePadding();
-    ImGui::TextUnformatted("Add:");
-    ImGui::SameLine(0.f, style.ItemInnerSpacing.x);
-    ImGui::SetNextItemWidth(-FLT_MIN);
-    const std::string preview = named ? fmt::format("{} {}", annotation_shape_icon(m_annotation_shape),
-                                                    annotation_shape_name(m_annotation_shape))
-                                      : annotation_shape_icon(m_annotation_shape);
-    if (ImGui::BeginCombo("##shape", preview.c_str()))
-    {
-        for (int n = 0; n < int(Annotation::Shape::COUNT); ++n)
-        {
-            const auto shape = Annotation::Shape(n);
-            // The names are always spelled out in the list, whatever the closed control has room for.
-            const bool is_selected = shape == m_annotation_shape;
-            if (ImGui::Selectable(
-                    fmt::format("{} {}", annotation_shape_icon(shape), annotation_shape_name(shape)).c_str(),
-                    is_selected))
-                m_annotation_shape = shape;
-            if (is_selected)
-                ImGui::SetItemDefaultFocus();
-        }
-        ImGui::EndCombo();
-    }
-    ImGui::SetItemTooltip("Which shape the annotate tool draws next.");
-}
-
-void HDRViewApp::draw_font_popup(Annotation &a, const ImVec2 &frame_padding, float item_spacing_x)
-{
-    if (!ImGui::BeginPopup("##font"))
-        return;
-
-    // The rows flatten their padding to stay one line tall; a popup off one of them is an ordinary window
-    // and wants ordinary widgets. Its own two rows sit closer than the default, which is spaced for a
-    // window.
-    ImGui::PushStyleVar(ImGuiStyleVar_FramePadding, frame_padding);
-    ImGui::PushStyleVar(ImGuiStyleVar_ItemSpacing, ImVec2(item_spacing_x, frame_padding.y));
-
-    const auto &faces = annotation_font_faces();
-    const char *shown = faces.front().label;
-    for (const auto &f : faces)
-        if (a.font_face == f.name)
-            shown = f.label;
-
-    ImGui::SetNextItemWidth(EmSize(8));
-    if (ImGui::BeginCombo("##face", shown))
-    {
-        for (const auto &f : faces)
-        {
-            const bool is_selected = a.font_face == f.name;
-            if (ImGui::Selectable(f.label, is_selected))
-            {
-                a.font_face                  = f.name;
-                m_annotation_style.font_face = a.font_face;
-            }
-            if (is_selected)
-                ImGui::SetItemDefaultFocus();
-        }
-        ImGui::EndCombo();
-    }
-
-    ImGui::SameLine(0.f, ImGui::GetStyle().ItemInnerSpacing.x);
-    ImGui::SetNextItemWidth(EmSize(6));
-    if (ImGui::DragFloat("##size", &a.font_size, 0.25f, Annotation::MinFontSize, Annotation::MaxFontSize, "%.0f px"))
-        m_annotation_style.font_size = a.font_size;
-    ImGui::SetItemTooltip("Image pixels, so the text grows and shrinks with what it labels.");
-
-    // Which corner or edge of the string lands on the point it was placed at, which is what NanoVG's
-    // align flags say. Each square carries a dot where its own anchor sits, drawn because FontAwesome has
-    // no diagonal arrows and four of the nine would be blank under it. Baseline is the tenth position and
-    // has no square; only a renderer sends it.
-    constexpr int h_of[3] = {VgCommand::AlignLeft, VgCommand::AlignCenter, VgCommand::AlignRight};
-    constexpr int v_of[3] = {VgCommand::AlignTop, VgCommand::AlignMiddle, VgCommand::AlignBottom};
-
-    static const char *const rows[3] = {"top", "middle", "bottom"};
-    static const char *const cols[3] = {"left", "center", "right"};
-
-    const float side = ImGui::GetFrameHeight() * 0.8f;
-    const float grid = 3.f * side + 2.f;
-
-    // Set against the middle of the grid. The offset goes inside the label's own group; outside it, it
-    // would move the first square alone, the other two lining up with the line the label left behind.
-    ImGui::BeginGroup();
-    ImGui::SetCursorPosY(ImGui::GetCursorPosY() + 0.5f * (grid - ImGui::GetTextLineHeight()));
-    ImGui::TextUnformatted("Anchor");
-    ImGui::EndGroup();
-    ImGui::SameLine(0.f, item_spacing_x);
-
-    ImGui::PushStyleVar(ImGuiStyleVar_ItemSpacing, ImVec2(1.f, 1.f));
-    ImGui::PushStyleVar(ImGuiStyleVar_FrameBorderSize, 1.f);
-    ImGui::BeginGroup();
-    for (int row = 0; row < 3; ++row)
-        for (int col = 0; col < 3; ++col)
-        {
-            ImGui::PushID(row * 3 + col);
-            if (col)
-                ImGui::SameLine();
-
-            const bool on = (a.text_align & h_of[col]) != 0 && (a.text_align & v_of[row]) != 0;
-            ImGui::PushStyleColor(ImGuiCol_Button, ImGui::GetColorU32(on ? ImGuiCol_ButtonActive : ImGuiCol_FrameBg));
-            if (ImGui::Button("##cell", ImVec2(side, side)))
-            {
-                a.text_align                  = h_of[col] | v_of[row];
-                m_annotation_style.text_align = a.text_align;
-            }
-            ImGui::PopStyleColor();
-            ImGui::SetItemTooltip("Anchor the text at its %s %s", rows[row], cols[col]);
-
-            // The dot sits where this square's own anchor would, so a square says what it does.
-            const ImVec2 lo = ImGui::GetItemRectMin(), hi = ImGui::GetItemRectMax();
-            const float  inset = 0.25f * side;
-            ImGui::GetWindowDrawList()->AddCircleFilled(ImVec2(ImLerp(lo.x + inset, hi.x - inset, 0.5f * float(col)),
-                                                               ImLerp(lo.y + inset, hi.y - inset, 0.5f * float(row))),
-                                                        ImMax(1.5f, side * 0.1f), ImGui::GetColorU32(ImGuiCol_Text));
-
-            ImGui::PopID();
-        }
-    ImGui::EndGroup();
-    ImGui::PopStyleVar(2);
-
-    ImGui::PopStyleVar(2);
-    ImGui::EndPopup();
-}
-
-void HDRViewApp::draw_annotation_controls(Annotation &a)
-{
-    const auto &style = ImGui::GetStyle();
-
-    // Icon alone, or icon and name once the row can spare the difference. It snaps between the two rather
-    // than growing with the panel, so the control does not change size as the panel is resized. Both widths
-    // leave room for the drop-down arrow, which is a frame wide at the right-hand end.
-    const float arrow_w   = ImGui::GetFrameHeight();
-    const float combo_pad = 2.f * style.FramePadding.x + arrow_w;
-
-    float names_w = 0.f;
-    for (int n = 0; n < int(Annotation::Shape::COUNT); ++n)
-        names_w = std::max(names_w, ImGui::CalcTextSize(annotation_shape_name(Annotation::Shape(n))).x);
-
-    const float combo_narrow = ImGui::IconSize().x + combo_pad;
-    const float combo_wide   = ImGui::IconSize().x + style.ItemInnerSpacing.x + names_w + combo_pad;
-
-    // "Add:" and the picker share a column, separated by their own spacing: in columns of their own the
-    // cell padding between them would set the label further from what it labels than from the drag beside.
-    const float add_text_w = ImGui::CalcTextSize("Add:").x + style.ItemInnerSpacing.x;
-    const float colors_w   = ImGui::GetFrameHeight();
-    const float width_min  = EmSize(3.5f);
-
-    // Spare width all goes to the width drag, until there is enough of it for the picker to spell its
-    // names out as well.
-    const float avail = ImGui::GetContentRegionAvail().x;
-    const bool  named = avail >= colors_w + width_min + add_text_w + combo_wide + 3.f * style.CellPadding.x;
-    const float add_w = add_text_w + (named ? combo_wide : combo_narrow);
-
-    if (ImGui::BeginTable("##AnnotationControls", 3,
-                          ImGuiTableFlags_NoSavedSettings | ImGuiTableFlags_NoPadOuterX |
-                              ImGuiTableFlags_SizingFixedFit))
-    {
-        ImGui::TableSetupColumn("add", ImGuiTableColumnFlags_WidthFixed, add_w);
-        ImGui::TableSetupColumn("colors", ImGuiTableColumnFlags_WidthFixed, colors_w);
-        ImGui::TableSetupColumn("width", ImGuiTableColumnFlags_WidthStretch, 1.f);
-        ImGui::TableNextRow();
-
-        draw_shape_picker(named);
-
-        ImGui::TableNextColumn();
-        bool restyled = stroke_fill_swatches("Colors", a.stroke_color, a.fill_color);
-
-        ImGui::TableNextColumn();
-        ImGui::SetNextItemWidth(-FLT_MIN);
-        restyled |= ImGui::DragFloat("##width", &a.stroke_width, 0.05f, 0.5f, 32.f, "%.1f px");
-        ImGui::SetItemTooltip("Stroke width in screen pixels, so it does not change with zoom.");
-
-        // Restyling the annotation in hand also sets what the next one will look like, so a color or a
-        // width chosen once carries forward instead of being forgotten when the selection is dropped.
-        if (restyled && &a != &m_annotation_style)
-        {
-            m_annotation_style.stroke_color = a.stroke_color;
-            m_annotation_style.fill_color   = a.fill_color;
-            m_annotation_style.stroke_width = a.stroke_width;
-        }
-
-        ImGui::EndTable();
-    }
-}
-
 std::vector<std::pair<int, int>> HDRViewApp::selected_targets() const
 {
     std::vector<std::pair<int, int>> out;
