+++ src/app-zoom.cpp
@@ -263,195 +263,6 @@ static float2 scroll_units(float2 wheel)
     return discrete ? wheel * k_units_per_notch : wheel;
 }
 
-int HDRViewApp::active_annotation() const
-{
-    auto img = current_image();
-    if (!img || m_active_annotation_on != img || m_active_annotation < 0 ||
-        m_active_annotation >= int(img->annotations.size()))
-        return -1;
-    return m_active_annotation;
-}
-
-void HDRViewApp::set_active_annotation(int index)
-{
-    m_active_annotation    = index;
-    m_active_annotation_on = index < 0 ? nullptr : current_image();
-}
-
-void HDRViewApp::cancel_annotation_drag()
-{
-    // Against the image the drag began on, which is not necessarily the one that is current now.
-    const auto held  = m_active_annotation_on;
-    const int  index = m_active_annotation;
-    if (held && index >= 0 && index < int(held->annotations.size()))
-    {
-        if (m_annotation_drag == AnnotationDrag::Creating)
-            held->annotations.erase(held->annotations.begin() + index);
-        else if (m_annotation_drag != AnnotationDrag::None)
-            held->annotations[size_t(index)] = m_annotation_drag_start;
-    }
-
-    if (m_annotation_drag == AnnotationDrag::Creating)
-        set_active_annotation(-1);
-
-    m_annotation_drag        = AnnotationDrag::None;
-    m_annotation_drag_handle = -1;
-}
-
-// How far the cursor must travel before a scribble keeps another point, and how far a point may stray from
-// its neighbors before being dropped again. Both screen pixels: a stroke should record the same shape
-// however far the image is zoomed.
-static constexpr float k_scribble_step = 2.f, k_scribble_tolerance = 1.f;
-
-void HDRViewApp::handle_annotate_tool()
-{
-    auto  img = current_image();
-    auto &io  = ImGui::GetIO();
-
-    // A drag whose image is no longer current is abandoned, half-drawn shape and all: playback or a close
-    // has put something else in its place.
-    if (m_annotation_drag != AnnotationDrag::None && m_active_annotation_on != img)
-    {
-        cancel_annotation_drag();
-        set_active_annotation(-1);
-    }
-
-    if (m_annotation_drag != AnnotationDrag::None && ImGui::IsKeyPressed(ImGuiKey_Escape))
-    {
-        cancel_annotation_drag();
-        return;
-    }
-
-    // Slop and handle radius are screen quantities, like everything else about how an annotation looks.
-    const float slop   = 0.25f * HelloImGui::EmSize();
-    const float radius = 0.3f * HelloImGui::EmSize();
-
-    // Screen pixels per image pixel, which turns the two scribble constants into the image distances the
-    // points are actually kept in.
-    const VgTransform xform_for_handles = viewport_transform();
-    const float       xform_scale       = xform_for_handles.scale;
-
-    auto        &list  = img->annotations;
-    const float2 pixel = pixel_at_app_pos(io.MousePos);
-
-    // Double-clicking a text annotation on the image edits what it says, which is where it is read.
-    if (ImGui::IsMouseDoubleClicked(ImGuiMouseButton_Left))
-        if (const int i = annotation_at(list, io.MousePos, viewport_transform(), slop);
-            i >= 0 && list[size_t(i)].shape == Annotation::Shape::Text)
-        {
-            set_active_annotation(i);
-            m_annotation_edit_text = true;
-
-            // The press that opened it had already taken hold of the annotation to move it.
-            m_annotation_drag = AnnotationDrag::None;
-            return;
-        }
-
-    if (ImGui::IsMouseClicked(ImGuiMouseButton_Left))
-    {
-        const auto xform  = viewport_transform();
-        const int  active = active_annotation();
-
-        // Handles of the annotation already in hand come first: they are drawn over everything, and often
-        // lie over a neighbor that would otherwise take the click.
-        if (active >= 0)
-            if (const int h = handle_at(list[size_t(active)], io.MousePos, xform, radius); h >= 0)
-            {
-                m_annotation_drag        = AnnotationDrag::Resizing;
-                m_annotation_drag_handle = h;
-                m_annotation_drag_start  = list[size_t(active)];
-                return;
-            }
-
-        if (const int i = annotation_at(list, io.MousePos, xform, slop); i >= 0)
-        {
-            set_active_annotation(i);
-            m_annotation_drag       = AnnotationDrag::Moving;
-            m_annotation_drag_start = list[size_t(i)];
-            m_annotation_grab       = pixel;
-            return;
-        }
-
-        // Empty space draws a new one, this being a drawing tool; the pan tool and a scroll still move the
-        // view. A click that never becomes a drag is undone on release, which is how clicking away clears
-        // the selection.
-        Annotation a = m_annotation_style;
-        a.shape      = m_annotation_shape;
-        a.points     = {pixel, pixel};
-        list.push_back(a);
-        set_active_annotation(int(list.size()) - 1);
-        m_annotation_drag       = AnnotationDrag::Creating;
-        m_annotation_drag_start = a;
-        return;
-    }
-
-    const int active = active_annotation();
-    if (m_annotation_drag == AnnotationDrag::None || active < 0)
-        return;
-
-    Annotation &a = list[size_t(active)];
-
-    if (ImGui::IsMouseDragging(ImGuiMouseButton_Left))
-    {
-        switch (m_annotation_drag)
-        {
-        case AnnotationDrag::Creating:
-            if (a.shape == Annotation::Shape::Freehand)
-            {
-                // A point per frame would record how fast the cursor moved, so one is kept only once the
-                // last is a couple of screen pixels behind.
-                const float step = k_scribble_step / std::max(xform_scale, 1e-6f);
-                if (length(pixel - a.points.back()) >= step)
-                    a.points.push_back(pixel);
-            }
-            else
-                a.p1() = pixel;
-            break;
-
-        case AnnotationDrag::Moving:
-        {
-            // Against the annotation as it was at mouse-down, so a move never accumulates rounding.
-            const float2 delta = pixel - m_annotation_grab;
-            a.points           = m_annotation_drag_start.points;
-            for (auto &p : a.points) p += delta;
-        }
-        break;
-
-        case AnnotationDrag::Resizing:
-            // Dragging a corner past its opposite renumbers the handles, so the drag follows the index
-            // back and keeps hold of the corner under the cursor.
-            m_annotation_drag_handle = move_annotation_handle(a, m_annotation_drag_handle, pixel, &xform_for_handles);
-            break;
-
-        default: break;
-        }
-    }
-    else if (ImGui::IsMouseReleased(ImGuiMouseButton_Left))
-    {
-        if (m_annotation_drag == AnnotationDrag::Creating)
-        {
-            // A captured scribble holds far more points than its shape needs. Simplified once, on
-            // release, so what is dropped is judged against the whole stroke.
-            if (a.shape == Annotation::Shape::Freehand)
-                a.points = simplify_polyline(a.points, k_scribble_tolerance / std::max(xform_scale, 1e-6f));
-
-            // A click that never became a drag leaves nothing behind: a shape with no extent is invisible
-            // and cannot be taken hold of again. Text is the exception, a click being how one is placed,
-            // and the panel opens its name for typing straight afterwards.
-            if (a.shape == Annotation::Shape::Text)
-                m_annotation_edit_text = true;
-            else if (a.bounds().min == a.bounds().max)
-            {
-                list.pop_back();
-                set_active_annotation(-1);
-            }
-        }
-
-        m_annotation_drag        = AnnotationDrag::None;
-        m_annotation_drag_handle = -1;
-    }
-}
-
 void HDRViewApp::handle_mouse_interaction()
 {
     auto &io = ImGui::GetIO();
